Move creation of `SolverConf` into `conf.rs`
